agents/bb42/requirements.txt
# Agent BB42 — onderwijshuisvesting (BBV 4.2)
requests>=2.31.0,<3
python-dotenv>=1.0.0,<2
